Cfgcpu: Implement remaining instructions. 16 bits tests passing. Logo working. Dune Hangs at some point

+++ src/Spice86.Core/Emulator/CPU/CfgCpu/ParsedInstruction/Instructions/Aaa.cs
@@ -4,7 +4,7 @@ using Spice86.Core.Emulator.CPU.CfgCpu.InstructionExecutor;
 using Spice86.Shared.Emulator.Memory;
 
 public class Aaa : CfgInstruction {
-    public Aaa(SegmentedAddress address, InstructionField<byte> opcodeField) :
+    public Aaa(SegmentedAddress address, InstructionField<ushort> opcodeField) :
         base(address, opcodeField) {
     }
