+++ src/Spice86.Core/Emulator/CPU/CfgCpu/ParsedInstruction/Instructions/CommonGrammar/InstructionWithSegmentRegisterIndexAndOffsetField.cs
@@ -5,12 +5,12 @@ using Spice86.Core.Emulator.CPU.CfgCpu.ParsedInstruction.Prefix;
 using Spice86.Shared.Emulator.Memory;
 
 public abstract class InstructionWithSegmentRegisterIndexAndOffsetField<T>: CfgInstruction, IInstructionWithSegmentRegisterIndex, IInstructionWithOffsetField<T> {
-    protected InstructionWithSegmentRegisterIndexAndOffsetField(SegmentedAddress address, InstructionField<byte> opcodeField, List<InstructionPrefix> prefixes, uint segmentRegisterIndex, InstructionField<T> offsetField) : base(address, opcodeField, prefixes) {
+    protected InstructionWithSegmentRegisterIndexAndOffsetField(SegmentedAddress address, InstructionField<ushort> opcodeField, List<InstructionPrefix> prefixes, int segmentRegisterIndex, InstructionField<T> offsetField) : base(address, opcodeField, prefixes) {
         SegmentRegisterIndex = segmentRegisterIndex;
         OffsetField = offsetField;
         FieldsInOrder.Add(offsetField);
     }
     
-    public uint SegmentRegisterIndex { get; }
+    public int SegmentRegisterIndex { get; }
     public InstructionField<T> OffsetField { get; }
 }
