+++ src/Spice86.Core/Emulator/CPU/CfgCpu/ParsedInstruction/Instructions/Hlt.cs
@@ -4,7 +4,7 @@ using Spice86.Core.Emulator.CPU.CfgCpu.InstructionExecutor;
 using Spice86.Shared.Emulator.Memory;
 
 public class Hlt : CfgInstruction {
-    public Hlt(SegmentedAddress address, InstructionField<byte> opcodeField) :
+    public Hlt(SegmentedAddress address, InstructionField<ushort> opcodeField) :
         base(address, opcodeField) {
     }
